What three audits found, and what was actually wrong
The most serious was in the quantity that looked most settled. `d` was logged
as the largest number of non-zeros in a column of the basis, following the
thesis's list of iteration-related measures. But `d` is also the divisor in the
normalisation stated four paragraphs later -- `||A_hat||_1 >= ||A_B||_1 /
(d ||A_B||max)` and `||A_hat||max >= 1/d` -- and those hold only when
`||A_B||_2 <= d ||A_B||max`. What is true is `||A||_2 <= sqrt(d_col d_row)
||A||max`, so the column maximum alone is too small whenever a row is denser
than every column, and then the two "lower bounds" are over-estimates. A
maximum-flow program does this at any vertex of degree four: on a thirty-relay
network, fifty-six of a hundred and twenty-one bases violate the bound and the
total lands sixteen per cent above what its own declared convention gives. `d`
is now the larger of the two readings -- which is also the sparsity of the
Hermitian dilation a quantum linear solver acts on, and what linsystem.py had
independently concluded. Both readings stay in the log.

Next most serious: a classical run that *failed* was costed as though it had
finished. `origin` special-cased only the literal lowercase "truncated", so a
log whose generated block said "failed" -- which an infeasible program writes,
records and all, and which `log --output` puts on disk -- came back EXACT with
no caveat anywhere, wearing the lemmas' LOWER label. A gate count for a solve
that did not happen is the exact failure this library exists around. It is
refused now, and the status is read case-insensitively so a hand-edited block
cannot lose its caveat to a capital letter.

Then the ones that were merely wrong. A one-arc DIMACS file declaring a billion
nodes was read as a billion-vertex network, because the node count was checked
for undercounting only -- isolated vertices being legitimate. It is not a wrong
number so much as an unusable one: the vertex count is the search space of every
quantum sweep, and Dinic tries to allocate an adjacency list per vertex first. An
MPS `FX` bound at 1e30 pinned a variable at 1e30, doing precisely what the
infinity convention exists to prevent. `_pattern_sparsity` accumulated an int8
matrix product in int8, so two constraint rows sharing 256 columns came out
structurally disjoint. The pivoting rule was honoured by the run and denied by
the provenance -- a Dantzig run logged and priced with the steepest-edge formula.
And QAA's displayed formula was built without the half its kernel carries, so
the printed expression evaluated to twice the printed number.

The CLI still handed out the broken template the previous commit claimed to have
fixed: the fix went into `dataset.template`, and `cli.py` was passing "csv"
explicitly. The new round-trip tests passed because they called the function, not
the command.

Much of the prose was stale in ways that matter. The field help for `kappa`,
`d`, `A_1` and `A_max` named the raw quantities while the instrumented run wrote
the normalised ones -- and that help is printed as the header of every generated
log, one line above the block saying the opposite, so a single file contradicted
itself and anyone logging their own GLPK run from it was silently costed under a
different scaling. The README's four published-comparison figures were exactly
twice the current output, having been written before the disputed half was
settled; they and the capability table are now regenerated from the code. web.py
claimed a test existed for the no-numpy promise, which it did not; the test
exists now, and so does its converse.

One audit finding is rejected. Lemma 8's prefactor is `24 sqrt(n-m) - 1`, as
coded: the proof derives it from `2^(n_c+1) - 1` with `n_c = log2(1/eps_QAE) +
log2(1 + 1/(2 delta_QAE))`, `eps_QAE = 1/(4 sqrt(n-m))` and `delta_QAE = 1/4`,
giving `2 * 12 sqrt(n-m) - 1`. The printed statement is mangled by OCR, not by
the transcription.

One is surfaced rather than settled, and it is the sort that needs the source:
`IPM/mnes` says it reuses the `QLS-Chebyshev` entry and does not. It imports
`chebyshev_terms`, never calls it, and inlines a degree with `log2(2/eps)` where
Lemma 16 has `log2(d kappa/eps)` -- a factor of 0.69 on the dominant term.
Either the paper's inner solver differs from Lemma 16, or one of the two is
wrong. It is written into CLAUDE.md as open.

+++ src/hybrid_benchmarking/classical/ipm.py
@@ -142,7 +142,11 @@ def _pattern_sparsity(matrix: np.ndarray) -> int:
     diagonal holds, and counting numerically would need a tolerance whose choice
     would then be a logged quantity in its own right.
     """
-    occupied = (matrix != 0).astype(np.int8)
+    # Counted in a width that cannot wrap. An int8 product accumulates in
+    # int8, so two rows sharing a multiple of 256 columns would come out as
+    # structurally disjoint -- and d enters the cycle count as gamma = d kappa
+    # with gamma squared inside it, so the understatement is not small.
+    occupied = (matrix != 0).astype(np.int64)
     return int(np.max(np.count_nonzero(occupied @ occupied.T, axis=1)))
